Fix encoding of special characters in the label
URL-encode all special characters (instead of only the colon and the
space character). This is necessary because otherwise the generated URL
is invalid if the label contains other special characters like for
example a question mark.

I think this is in alignment with what the specification says: "The
label is used to identify which account a key is associated with. It
contains an account name, which is a URI-encoded string, optionally
prefixed by an issuer string identifying the provider or service
managing that account."

https://github.com/google/google-authenticator/wiki/Key-Uri-Format